requirements.txt
numpy==2.2.6
opencv-python==4.12.0.88
python-dotenv==1.1.1
PyYAML==6.0.2
